chore: consolidate browser utils
