+++ packages/instrumentation/src/browser-utils/index.ts
