+++ packages/instrumentation/src/browser-utils/testing.test.ts
@@ -5,7 +5,7 @@
 
 import { logs } from '@opentelemetry/api-logs';
 import { describe, expect, it } from 'vitest';
-import { setupTestLogExporter } from './logs.ts';
+import { setupTestLogExporter } from './testing.ts';
 
 describe('logTestUtils', () => {
   it('should collect logs in memory', () => {
