--- packages/instrumentation/src/test-utils/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { setupTestLogExporter } from './logs.ts';
