+++ packages/instrumentation/src/web-vitals/instrumentation.test.ts
@@ -14,7 +14,7 @@ import {
   vi,
 } from 'vitest';
 import { page, userEvent } from 'vitest/browser';
-import { setupTestLogExporter } from '#instrumentation-test-utils';
+import { setupTestLogExporter } from '#browser-test-utils';
 import { WebVitalsInstrumentation } from './instrumentation.ts';
 import {
   ATTR_WEB_VITAL_DELTA,
