+++ packages/instrumentation/tsdown.config.ts
@@ -3,5 +3,5 @@ import baseConfig from '../../tsdown.config.ts';
 
 export default defineConfig({
   ...baseConfig,
-  entry: ['src/*/index.ts', '!src/{test-utils,utils}/index.ts'],
+  entry: ['src/*/index.ts', '!src/browser-utils/index.ts'],
 });
